Fix all but one remaining warnings, and some bugs
Most of the warnings and all of the bugs are to do with printf formats.
Hooray for GCC's attributes that allow user functions' formats to be
checked!

+++ src/completion.c
@@ -163,7 +163,7 @@ write_completion (va_list ap)
   bprintf ("Possible completions are:\n");
   for (size_t i = 0, col = 0; i < gl_list_size (l); i++)
     {
-      bprintf ("%-*s", max, (const char *) gl_list_get_at (l, i));
+      bprintf ("%-*s", (int) max, (const char *) gl_list_get_at (l, i));
 
       col = (col + 1) % numcols;
       if (col == 0)
